pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-blackbox-arena"
version = "0.1.0"
description = "OpenEnv-style repair gym for agent reliability CI."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
]

[project.optional-dependencies]
training = [
  "trl>=0.15",
  "transformers>=4.48",
  "accelerate>=0.34",
  "datasets>=2.20",
  "peft>=0.12",
]
plots = [
  "pillow>=8",
]
dev = [
  "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_blackbox*", "server*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
